Organizes Composition layer components into separate files.
