+++ FeedModule/FeediOS/Feed UI/Composers/FeedLoadingPresentationAdapter.swift
@@ -1,0 +1,29 @@
+//
+//  FeedLoadingPresentationAdapter.swift
+//  FeediOS
+//
+//  Created by Omran Khoja on 2/22/22.
+//
+
+import FeedModule
+
+final class FeedLoadingPresentationAdapter: FeedViewControllerDelegate {
+    private let feedLoader: FeedLoader
+    var presenter: FeedPresenter?
+    
+    init(feedLoader: FeedLoader) {
+        self.feedLoader = feedLoader
+    }
+    
+    func didRequestFeedRefresh() {
+        presenter?.didStartLoadingFeed()
+        feedLoader.load { [weak self] result in
+            switch result {
+            case let .success(feed):
+                self?.presenter?.didFinishLoadingFeed(feed: feed)
+            case let .failure(error):
+                self?.presenter?.didFinishLoadingFeed(with: error)
+            }
+        }
+    }
+}
