+++ FeedModule/FeediOS/Feed UI/Composers/FeedViewAdapter.swift
@@ -1,0 +1,27 @@
+//
+//  FeedViewAdapter.swift
+//  FeediOS
+//
+//  Created by Omran Khoja on 2/22/22.
+//
+
+import UIKit
+
+final class FeedViewAdapter: FeedView {
+    private weak var feedController: FeedViewController?
+    private let imageLoader: FeedImageDataLoader
+    
+    init(feedController: FeedViewController, imageLoader: FeedImageDataLoader){
+        self.feedController = feedController
+        self.imageLoader = imageLoader
+    }
+    
+    func display(_ viewModel: FeedViewModel) {
+        feedController?.viewModel = viewModel.feed.map { model in
+            let adapter = ImageLoadingPresentationAdapter<WeakRefVirtualProxy<FeedImageCellController>, UIImage>(model: model, imageLoader: imageLoader)
+            let view = FeedImageCellController(delegate: adapter)
+            adapter.presenter = FeedImagePresenter(view: WeakRefVirtualProxy(view), imageTransformer: UIImage.init)
+            return view
+        }
+    }
+}
